stream_body uses response.body = iostream
instead of a loop of response.stream.write's.

Rails still streams like this. But it is delivering the HTTP response
headers a lot faster this way, almost immediately.

With the previous way, with a 1G original file accessed via
/downloads, I was seeing an 11 second wait even to get HTTP headers.

Not sure why, it could be because it takes that long to even get
the first byte from fedora, and it's not sending anything until it does?
But I tried to mess around with things using the response.stream.write
API, and was not able to improve things.

Browsers are willing to wait that long, but it's not a great UX -- you get
a 'waiting for connection' poup in chrome until those original headers, at
which point it shows up as a download in progress.

But some non-browser user agents are not willing to wait that long. imgix.com
is willing to wait at most 10 seconds for HTTP headeres (after that it will wait quite a bit longer for the entire content to stream). Other CDNs may be similar.

Ensuring quick HTTP header response seems a good idea, and this change
seems to accomplish it. Even my 1G file had nearly immediately HTTTP header
delivery, although of course you still have to wait for the file.

For some background on response.body = stream as a valid rails
streaming API, see https://github.com/rails/rails/issues/18714