Fix download streaming with self.response_body=
We have been seeing lots of extra memory usage when downloading large
files from Fedora. There was a version of this fix applied in #421, but
it was reverted in #427. This uses the recommended technique of setting
the controller's response_body to the stream.

In looking through the ActionController source, this should both start
the output (sending headers), and flag the render as having performed,
so implicit template resolution/rendering will be skipped.